Add callbacks to sqlalchemy events.
We encountered a use case where we want to show the parameters
of our sql query for better traceability. I added a callback parameter
to the EngineTracer in order to capture the query and the span's context
for further processing.

I also noticed that if the engine is instrumented through the
connect wrapper, the engine tracer options such as the commenter
won't be configured. I packed these options together so users
can pass the configuration regardless of which wrapper is being used.